--- src/main/java/Administrator/AdministratorManagementServices/UserManagement.java
@@ -1,11 +1,0 @@
-package Administrator.AdministratorManagementServices;
-import User.user;
-
-
-public interface UserManagement {
-
-    public boolean addCustomer (user user)     ;
-    public boolean updateCustomer (user user)  ;
-    public boolean deleteCustomer (user user)  ;
-    public void handleCustomerManagement(int userChoice);
-}
