+++ src/main/java/Administrator/Managers/ProductService/ProductManagement.java
@@ -1,4 +1,4 @@
-package Administrator.AdministratorManagementServices;
+package Administrator.Managers.ProductService;
 
 import Classes.Products.Product;
