+++ src/main/java/Classes/InterfaceManagers/Interface.java
@@ -2,6 +2,5 @@ package Classes.InterfaceManagers;
 
 public abstract class Interface {
 
-    public void printInterface(){}
     public void printUserFunctions(){}
 }
