+++ src/main/java/Supplier/CustomerInterfaceManager.java
@@ -3,10 +3,7 @@ package Supplier;
 import Classes.InterfaceManagers.Interface;
 
 public class CustomerInterfaceManager extends Interface {
-    @Override
-    public void printInterface() {
 
-    }
 
     @Override
     public void printUserFunctions() {
