+++ src/main/java/UserDAO/UserDAOInterface.java
@@ -1,0 +1,8 @@
+package UserDAO;
+
+import UserClass.User;
+
+public interface UserDAOInterface {
+
+    public void storeUser(User user);
+}
