+++ src/main/java/UserDAO/UserDAOManager.java
@@ -1,0 +1,10 @@
+package UserDAO;
+
+import UserClass.User;
+
+public class UserDAOManager implements UserDAOInterface{
+    @Override
+    public void storeUser(User user) {
+        user.storeDB(user);
+    }
+}
